Finished lab - 100 /100

+++ 06. Data types and variables/01_CenturiesToMinutes/Program.cs
@@ -7,10 +7,11 @@ namespace _01_CenturiesToMinutes
         static void Main(string[] args)
         {
             int centuries = int.Parse(Console.ReadLine());
-            int years = centuries * 100;
-            int days = (int)(years * 365.2422);
-            int hours = days * 24;
-            int minutes = hours * 60;
+
+            long years = centuries * 100;
+            long days = (int)(years * 365.2422);
+            long hours = days * 24;
+            long minutes = hours * 60;
 
             Console.WriteLine("{0} centuries = {1} years = {2} days = {3} hours = {4} minutes", 
                 centuries, years, days, hours, minutes);
